+++ demos/world/src/main/java/is/codion/demos/world/domain/WorldImpl.java
@@ -253,7 +253,8 @@ public final class WorldImpl extends DomainModel {
 										// end::booleanColumn[]
 										// tag::derivedAttribute[]
 										CountryLanguage.NO_OF_SPEAKERS.define()
-														.derived(CountryLanguage.COUNTRY_FK, CountryLanguage.PERCENTAGE)
+														.derived()
+														.from(CountryLanguage.COUNTRY_FK, CountryLanguage.PERCENTAGE)
 														.value(new NoOfSpeakers())
 														.caption("No. of speakers")
 														.numberGrouping(true),