+++ framework/domain/src/test/java/is/codion/framework/domain/TestDomain.java
@@ -256,7 +256,8 @@ public final class TestDomain extends DomainModel {
 														.items(INT_VALUE_ITEMS)
 														.caption(Detail2.INT_ITEMS.name()),
 										Detail2.INT_DERIVED.define()
-														.derived(Detail2.INT)
+														.derived()
+														.from(Detail2.INT)
 														.value(source -> {
 															Integer intValue = source.get(Detail2.INT);
 															if (intValue == null) {
@@ -363,7 +364,8 @@ public final class TestDomain extends DomainModel {
 														.items(INT_VALUE_ITEMS)
 														.caption(Detail.INT_ITEMS.name()),
 										Detail.INT_DERIVED.define()
-														.derived(Detail.INT)
+														.derived()
+														.from(Detail.INT)
 														.value(source -> {
 															Integer intValue = source.get(Detail.INT);
 															if (intValue == null) {
@@ -512,7 +514,8 @@ public final class TestDomain extends DomainModel {
 														.denormalized(Employee.DEPARTMENT_FK, Department.LOCATION)
 														.caption(Department.LOCATION.name()),
 										Employee.DEPARTMENT_NAME.define()
-														.derived(Employee.NAME, Employee.DEPARTMENT_FK)
+														.derived()
+														.from(Employee.NAME, Employee.DEPARTMENT_FK)
 														.value(new DepartmentName()),
 										Employee.DATA.define()
 														.column()
@@ -650,7 +653,8 @@ public final class TestDomain extends DomainModel {
 										InvalidDerived.INT.define()
 														.column(),
 										InvalidDerived.INVALID_DERIVED.define()
-														.derived(InvalidDerived.ID)
+														.derived()
+														.from(InvalidDerived.ID)
 														.value(source -> source.get(InvalidDerived.INT).intValue()))
 						.caption(InvalidDerived.INVALID_DERIVED.name())//incorrect source value, trigger exception
 						.formatter(entity -> null)