+++ changelog.md
@@ -4,6 +4,8 @@ Codion Change Log
 ## 0.18.51
 ### is.codion.swing.common.ui
 - FilterTable.Builder.hiddenColumns() added.
+### is.codion.framework.domain
+- DerivedAttributeDefinition.Builder.SourceAttributesStep added.
 ### is.codion.swing.framework.ui
 - EntitySearchFieldPanel.Builder.preferredSearchFieldWidth() added.
 - EntityPanel.createMainComponent() bug fixed, mainPanel() got called twice if a detail panel was present, messing up the controls panel, now supplier based.